Add Image Selection Chevrons
Chevrons only display when there are more than nine images. When
clicked, they slide images into view left or right depending on which
chevron is clicked. This is achieved by updating that translate property
on the image-container element.

When the main image chevrons are clicked so that an image is selected
is not currently in view, that image is slid into view in the
image-container element.

Tests are included.